research_agent/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "arxiv>=2.4.0",
    "fastapi[standard]>=0.128.3",
    "langchain>=1.2.9",
    "langchain-arxiv>=0.1.0",
    "langchain-community>=0.4.1",
    "pymysql>=1.1.2",
    "snowflake-id>=1.0.2",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]

[[tool.uv.index]]
url = "http://mirrors.aliyun.com/pypi/simple"
default = true
